requirements.txt
python-dotenv~=0.13.0
python-lichess~=0.9
python-chess~=0.31.2
jupyterlab~=2.1.4
pytest~=5.4.3
pdoc3~=0.8.1
pylint~=2.5.3
pgnparser~=1.0
coverage~=5.1
pytest-cov~=2.10.0
mypy~=0.780
setuptools~=47.1.1